==> 05_eid_to_osmid_mappings.ipynb <==
import sys
import geopandas as gpd
display(sys.version)
display(gpd.__version__)

df = gpd.read_file('porto/edges.shp')

edge_eid_to_osmid = df[['osmid0']].rename(columns={'osmid0': 'osmid'})
edge_eid_to_osmid.to_csv('data/edges_eid_to_osmid.csv', index=False)
edge_eid_to_osmid.head(5)
